refactor: extract shared done-gate test harness
writeFeatureTicketAtDone + the stop-quality.ts runner were about to be
duplicated between skill-gate-integration.test.ts and the new Codex/Cursor
fallback E2E test. Extract them into done-gate-harness.ts; runDoneGate takes
an optional env so the Codex test can pass a CLAUDE_SESSION_ID-scrubbed
environment. codex-cursor-skill-fallback.test.ts now imports the harness.

No behavior change — 8 tests green.
